+++ packages/core-components/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @evidence-dev/core-components
 
+## 2.0.0-usql.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @evidence-dev/component-utilities@2.0.0-usql.4
+
+## 2.0.0-usql.5
+
+### Patch Changes
+
+- Updated dependencies [64ab3074]
+  - @evidence-dev/component-utilities@2.0.0-usql.3
+
+## 2.0.0-usql.4
+
+### Patch Changes
+
+- 9ade9c88: Add Definitions component
+- 9432c6e4: limit `getFormatObjectFromString` in Value.svelte
+- be1cc666: force NaN, null, and undefined to the top of ascending sort order and vice versa
+
+## 1.2.0
+
+### Minor Changes
+
+- 78f2fab2: Adds modal, accordion and link button UI components
+
+### Patch Changes
+
+- 75560a31: Consolidate tailwind presets into tailwind package
+
+## 2.0.0-usql.3
+
+### Patch Changes
+
+- Updated dependencies [e1174aa1]
+  - @evidence-dev/component-utilities@2.0.0-usql.2
+
+## 2.0.0-usql.2
+
+### Minor Changes
+
+- 78f2fab2: Adds modal, accordion and link button UI components
+
 ## 2.0.0-usql.1
 
 ### Patch Changes