+++ packages/csv/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @evidence-dev/csv
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies [4e783f36]
+- Updated dependencies [e12fef6c]
+  - @evidence-dev/duckdb@0.1.1
+
+## 1.0.0-usql.2
+
+### Minor Changes
+
+- 3b49d3b6: Update csv adapter to function as db plugin
+
+## 1.0.0-usql.1
+
+### Patch Changes
+
+- Updated dependencies [e12fef6c]
+  - @evidence-dev/duckdb@1.0.0-usql.1
+
 ## 1.0.0-usql.0
 
 ### Major Changes