+++ packages/plugin-connector/CHANGELOG.md
@@ -1,5 +1,61 @@
 # @evidence-dev/plugin-connector
 
+## 2.0.0-usql.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @evidence-dev/universal-sql@2.0.0-usql.6
+
+## 2.0.0-usql.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @evidence-dev/universal-sql@2.0.0-usql.5
+
+## 2.0.0-usql.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @evidence-dev/universal-sql@2.0.0-usql.4
+
+## 2.0.0-usql.9
+
+### Minor Changes
+
+- cfb0f248: Respect component plugin's tailwind configuration
+
+### Patch Changes
+
+- 8ffbb361: skip massive files and non-directories in sources
+
+## 2.0.0-usql.8
+
+### Minor Changes
+
+- e6f550f3: Improve data source return type validation to be more performance and stringent
+
+## 2.0.0-usql.7
+
+### Patch Changes
+
+- Updated dependencies [ca7337ba]
+  - @evidence-dev/universal-sql@2.0.0-usql.3
+
+## 2.0.0-usql.5
+
+### Patch Changes
+
+- df7a8c5a: Ignore source files larger than 100Mb
+
+## 2.0.0-usql.4
+
+### Patch Changes
+
+- cff22ece: Only read files on demand, prevents attempted loading of very large db files
+
 ## 2.0.0-usql.3
 
 ### Patch Changes